+++ docs/.vuepress/components/Bit.vue
@@ -1,8 +1,0 @@
-<template>
-  <p class="bit-sponsor">
-    <a href="https://bit.dev/?utm_source=vue&utm_medium=vue&utm_campaign=vue&utm_term=vue&utm_content=vue" target="_blank">
-      <span>This project is sponsored by</span>
-      <img alt="bit" src="/bit-wide.png">
-    </a>
-  </p>
-</template>