+++ examples/server.js
@@ -27,7 +27,8 @@ fs.readdirSync(__dirname).forEach(file => {
 
 app.use(express.static(__dirname))
 
+const host = process.env.HOST || 'localhost'
 const port = process.env.PORT || 8080
-module.exports = app.listen(port, () => {
-  console.log(`Server listening on http://localhost:${port}, Ctrl+C to stop`)
+module.exports = app.listen(port, host, () => {
+  console.log(`Server listening on http://${host}:${port}, Ctrl+C to stop`)
 })