+++ test/e2e/specs/active-links.js
@@ -5,6 +5,7 @@ module.exports = {
 
   '@tags': ['history', 'active', 'router-link'],
 
+  /** @type {import('nightwatch').NightwatchTest} */
   'active links': function (browser) {
     browser
       .url('http://localhost:8080/active-links/')
@@ -31,6 +32,7 @@ module.exports = {
       .assert.attributeContains('li:nth-child(18) a', 'href', '/active-links/redirect-image')
       .assert.attributeContains('li:nth-child(19) a', 'href', '/active-links/redirect-image')
       .assert.containsText('.view', 'Home')
+      .assert.not.attributeEquals(`li:nth-child(3) a`, 'aria-current', 'page')
 
     assertActiveLinks(1, [1, 2], null, [1, 2])
     assertActiveLinks(2, [1, 2], null, [1, 2])
@@ -70,12 +72,14 @@ module.exports = {
         browser.assert
           .cssClassPresent(`li:nth-child(${i}) a`, 'router-link-exact-active')
           .assert.cssClassPresent(`li:nth-child(${i}) a`, 'router-link-active')
+          .assert.attributeEquals(`li:nth-child(${i}) a`, 'aria-current', 'page')
       })
       exactActiveLI &&
         exactActiveLI.forEach(i => {
           browser.assert
             .cssClassPresent(`li:nth-child(${i})`, 'router-link-exact-active')
             .assert.cssClassPresent(`li:nth-child(${i})`, 'router-link-active')
+            .assert.attributeEquals(`li:nth-child(${i}) a`, 'aria-current', 'page')
         })
     }
   }